+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 
 - HCL lib version from V1 to V2 and all the implications
   ([PR #135](https://github.com/cycloidio/terracognita/pull/135))
+- All the Provider and Terraform versions
+  ([PR #143](https://github.com/cycloidio/terracognita/pull/143))
 
 ### Fixed